+++ R/00-Namespace.R
@@ -3,6 +3,7 @@
 #' @importFrom coda raftery.diag gelman.diag geweke.diag heidel.diag mcmc
 #' @importFrom colourpicker colourInput
 #' @importFrom rcarbon calibrate sampleDates
+#' @importFrom dplyr bind_rows everything group_by group_indices inner_join left_join select
 #' @importFrom DT datatable renderDataTable dataTableOutput
 #' @importFrom geometry convhulln inhulln
 #' @importFrom ggplot2 ggplot theme theme_light coord_cartesian geom_point theme_light theme labs
@@ -28,6 +29,7 @@
 #' @importFrom readr guess_encoding
 #' @importFrom readxl excel_sheets read_excel
 #' @importFrom Rfast spdinv Crossprod rmvnorm
+#' @importFrom rlang .data
 #' @importFrom shinyjs alert reset runjs useShinyjs
 #' @importFrom shinyMatrix matrixInput updateMatrixInput
 #' @importFrom shinyWidgets pickerInput updatePickerInput
@@ -37,6 +39,8 @@
 #' @importFrom stats aggregate as.formula model.matrix na.exclude na.omit predict quantile dnorm
 #'  qnorm residuals pnorm var sd setNames cor density median kmeans
 #' @importFrom stringi stri_escape_unicode
+#' @importFrom templates tmpl
+#' @importFrom tidyr separate unite
 #' @importFrom utils available.packages compareVersion install.packages head
 #'  packageVersion read.csv read.csv2 write.table installed.packages capture.output download.file
 #' @importFrom zip zipr